requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

# Project 1: Medical Multimodal Report Generation
# Core Deep Learning
torch==2.7.0+cu128
torchvision==0.22.0+cu128
pytorch-lightning>=2.0.0

# Transformers & NLP
transformers>=4.35.0
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.3.0
nibabel>=5.1.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Metrics & Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
pycocoevalcap>=1.2

# Model Serving
gradio>=4.0.0
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
